refactored; Event lat/long to BigDecimal; EventParticipant with ManyToOne + Serializable ID; EventRepository find by organizer; added; out-of-order migrations; recreated; migration scripts for participants + event table.

+++ src/main/java/com/webapp/Eventified/service/EventService.java
@@ -1,5 +1,6 @@
 package com.webapp.Eventified.service;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -19,7 +20,7 @@ public class EventService {
     private final EventRepository eventRepository;
     private final UserRepository userRepository;
 
-    public Event createEvent(UUID userId, String title, Integer sport, String address, Integer skillLevel, LocalDateTime startTime, LocalDateTime endTime, Integer capacity, Double latitude, Double longitude) {
+    public Event createEvent(UUID userId, String title, Integer sport, String address, Integer skillLevel, LocalDateTime startTime, LocalDateTime endTime, Integer capacity, BigDecimal latitude, BigDecimal longitude) {
 
         if (eventRepository.findByTitleAndOrganizer_Id(title, userId).isPresent()){ 
             throw new IllegalArgumentException("Event with the same title already exists for this user.");
